feat(frontend): modulos de Inventario/Kardex y Reportes consumiendo la API
Inventario / Kardex (solo lectura):
- StockMovementService + lista con filtros (producto, tipo, rango de fechas),
  badges por tipo, delta con signo y stock antes/despues, paginacion

Reportes (solo Administrador):
- ReportService + pagina con 3 vistas (Ventas, Top productos, Inventario)
  via tabs; rango de fechas, orden y limite
- Componente reutilizable BarChart (CSS puro, sin dependencias) para los graficos
- Stat cards de totales (ingresos, ISV, descuentos, promedio diario, etc.)

Backend: cast del parametro limit a int en reports/products (evita TypeError
con strict_types al recibirlo del query string).

Reemplaza los placeholders ComingSoon de /stock-movements y /reports.
styles.scss: utilidades de tabs, stat cards, panel, filtro de fecha y deltas.

+++ pharmacy-app/app/Http/Controllers/Api/ReportController.php
@@ -37,7 +37,7 @@ class ReportController extends Controller
         $top = $reports->topProducts(
             $from,
             $to,
-            $validated['limit'] ?? 10,
+            (int) ($validated['limit'] ?? 10),
             $validated['sort_by'] ?? 'quantity',
         );
